Documentation + Warnings

+++ Statutis.API/Controllers/GroupController.cs
@@ -1,17 +1,18 @@
-using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Statutis.API.Form;
 using Statutis.API.Models;
 using Statutis.Core.Interfaces.Business;
 using Statutis.Core.Interfaces.Business.History;
 using Statutis.Core.Interfaces.Business.Service;
-using Statutis.Entity;
 using Statutis.Entity.Service;
 
 namespace Statutis.API.Controllers;
 
+/// <summary>
+/// Controleur sur les goupes
+/// </summary>
+[Tags("Groupe")]
 [Route("api/groups")]
 [ApiController]
 public class GroupController : Controller
@@ -21,6 +22,13 @@ public class GroupController : Controller
 	private readonly IUserService _userService;
 	private readonly ITeamService _teamService;
 
+	/// <summary>
+	/// Constructeur
+	/// </summary>
+	/// <param name="historyEntryService"></param>
+	/// <param name="groupService"></param>
+	/// <param name="userService"></param>
+	/// <param name="teamService"></param>
 	public GroupController(IHistoryEntryService historyEntryService, IGroupService groupService, IUserService userService, ITeamService teamService)
 	{
 		_historyEntryService = historyEntryService;
@@ -29,6 +37,10 @@ public class GroupController : Controller
 		this._teamService = teamService;
 	}
 
+	/// <summary>
+	/// Récupration de tous les groupes
+	/// </summary>
+	/// <returns>Groupes public si pas authentifié, sinon tous les groupes existants</returns>
 	[HttpGet, Route("")]
 	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<GroupModel>))]
 	public async Task<IActionResult> GetAll()
@@ -38,7 +50,8 @@ public class GroupController : Controller
 		if (HttpContext.User.Identity?.IsAuthenticated ?? false)
 		{
 			var user = await _userService.GetUserAsync(User);
-			res.AddRange(await _groupService.GetFromUser(user));
+			if (user != null)
+				res.AddRange(await _groupService.GetFromUser(user));
 		}
 
 		return Ok(
@@ -48,9 +61,16 @@ public class GroupController : Controller
 		);
 	}
 
+	/// <summary>
+	/// Récupération d'un groupe par son identifiant
+	/// </summary>
+	/// <param name="guid">Identifiant du groupe cible</param>
+	/// <returns>Un groupe</returns>
+	/// <response code="403">Si vous ne diposez pas des accès suivant pour consulter ces informations.</response>
 	[HttpGet, Route("{guid}")]
 	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(GroupModel))]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
+	[ProducesResponseType(StatusCodes.Status403Forbidden)]
 	public async Task<IActionResult> Get(Guid guid)
 	{
 		Group? group = await _groupService.Get(guid);
@@ -64,6 +84,9 @@ public class GroupController : Controller
 
 
 			var user = await _userService.GetUserAsync(User);
+			if (user == null)
+				return Forbid();
+
 			var teamUserId = user.Teams.Select(x => x.TeamId);
 			if (!group.Teams.Any(x => teamUserId.Contains(x.TeamId)))
 				return Forbid();
@@ -74,9 +97,19 @@ public class GroupController : Controller
 		return Ok(new GroupModel(group, histories, Url));
 	}
 
+	/// <summary>
+	///	Mise a jour d'un groupe
+	/// </summary>
+	/// <param name="guid">Identifiant du groupe que l'on souhaite mettre à jour</param>
+	/// <param name="form">Nouvelles informations sur le groupe</param>
+	/// <returns></returns>
+	/// <response code="401">Si vous n'êtes pas authentifié.</response>
+	/// <response code="403">Si vous ne diposez pas des accès suivant pour consulter ces informations.</response>
 	[HttpPut, Route("{guid}"), Authorize]
 	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(GroupModel))]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
+	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
+	[ProducesResponseType(StatusCodes.Status403Forbidden)]
 	public async Task<IActionResult> Update(Guid guid, [FromBody] GroupForm form)
 	{
 		Group? group = await _groupService.Get(guid);
@@ -89,6 +122,9 @@ public class GroupController : Controller
 
 
 		var user = await _userService.GetUserAsync(User);
+		if (user == null)
+			return Forbid();
+
 		var teamUserId = user.Teams.Select(x => x.TeamId);
 		if (!group.Teams.Any(x => teamUserId.Contains(x.TeamId)))
 			return Forbid();
@@ -114,9 +150,15 @@ public class GroupController : Controller
 
 	}
 
+	/// <summary>
+	///	Ajout d'un groupe
+	/// </summary>
+	/// <param name="form">Informations sur le nouveau groupe</param>
+	/// <returns></returns>
+	/// <response code="401">Si vous n'êtes pas authentifié.</response>
 	[HttpPost, Route(""), Authorize]
 	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(GroupModel))]
-	[ProducesResponseType(StatusCodes.Status404NotFound)]
+	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
 	public async Task<IActionResult> Add([FromBody] GroupForm form)
 	{
 		Group group = new Group(form.Name, form.Description);
@@ -137,10 +179,18 @@ public class GroupController : Controller
 		return Ok(new GroupModel(group, histories, Url));
 
 	}
-	
+
+	/// <summary>
+	///	Suppression d'un groupe
+	/// </summary>
+	/// <param name="guid">Identifiant du groupe que l'on souhaite supprimer</param>
+	/// <returns></returns>
+	/// <response code="401">Si vous n'êtes pas authentifié.</response>
+	/// <response code="403">Si vous ne diposez pas des accès suivant pour consulter ces informations.</response>
 	[HttpDelete, Route("{guid}"), Authorize]
 	[ProducesResponseType(StatusCodes.Status200OK)]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
+	[ProducesResponseType(StatusCodes.Status403Forbidden)]
 	public async Task<IActionResult> Delete(Guid guid)
 	{
 		Group? group = await _groupService.Get(guid);
@@ -153,23 +203,35 @@ public class GroupController : Controller
 
 
 		var user = await _userService.GetUserAsync(User);
+		if (user == null)
+			return Forbid();
+
 		var teamUserId = user.Teams.Select(x => x.TeamId);
 		if (!group.Teams.Any(x => teamUserId.Contains(x.TeamId)))
 			return Forbid();
 
 
 		await _groupService.Delete(group);
-			
-		
+
+
 		return Ok();
 
 	}
-	
+
+	/// <summary>
+	/// Récupération d'un avatar d'un groupe
+	/// </summary>
+	/// <param name="guid">Identifiant du groupe cible</param>
+	/// <remarks>Lorsque l'on est pas authentifié, il n'est possible de récupérer que les avatars des groupes publics</remarks>
+	/// <returns></returns>
+	/// <response code="404">Si le groupe visé n'existe pas ou qu'il ne dispose pas d'avatar.</response>
+	/// <response code="403">Vous ne disposez pas des droits suffisants.</response>
 	[HttpGet]
 	[AllowAnonymous]
 	[Route("avatar/{guid}")]
-	[ProducesResponseType(StatusCodes.Status200OK)]
+	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(FileContentResult))]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
+	[ProducesResponseType(StatusCodes.Status403Forbidden)]
 	public async Task<IActionResult> GetAvatar(Guid guid)
 	{
 		var targetGroup = await _groupService.Get(guid);
@@ -181,11 +243,23 @@ public class GroupController : Controller
 		return File(targetGroup.Avatar, targetGroup.AvatarContentType);
 	}
 
+	/// <summary>
+	/// Mettre à jour l'avatar d'un groupe
+	/// </summary>
+	/// <param name="guid">Identifiant du groupe cible</param>
+	/// <param name="form">Informations sur le nouvel avatar (null si l'on souhaite supprimer celui courant)</param>
+	/// <returns></returns>
+	/// 
+	/// <remarks>Il n'est possible de modifier l'avatar que si l'utilisateur à un accès</remarks>
+	/// <response code="404">Si le groupe visé n'existe pas.</response>
+	/// <response code="401">Si vous n'êtes pas authentifié.</response>
+	/// <response code="403">Vous ne disposez pas des droits suffisants.</response>
 	[HttpPut, Authorize]
 	[Authorize]
 	[Route("avatar/{guid}")]
 	[ProducesResponseType(StatusCodes.Status200OK)]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
+	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
 	public async Task<IActionResult> UploadAvatar(Guid guid, IFormFile? form = null)
 	{
 
@@ -195,9 +269,9 @@ public class GroupController : Controller
 
 		var targetGroup = await _groupService.Get(guid);
 		var userGroups = await _groupService.GetFromUser(user);
-		
+
 		if (targetGroup == null || (user.Roles != "ROLE_ADMIN" && userGroups.Contains(targetGroup)))
-			return NotFound();
+			return Forbid();
 
 
 		if (form == null)
@@ -220,6 +294,6 @@ public class GroupController : Controller
 
 		return Ok();
 	}
-	
+
 
 }
